disease-detection/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
numpy==2.0.1
scikit-learn==1.6.1
joblib==1.4.2
pymongo==4.6.3
certifi==2025.1.31
opencv-python==4.8.1.78
requests==2.31.0
python-dotenv==1.0.0